requirements.txt
# Machine Learning Coursework 2 - Image Classification
# Required Python packages

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Data Science and Numerical Computing
numpy>=1.24.0
pandas>=2.0.0

# Machine Learning
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Image Processing
Pillow>=10.0.0

# Progress bars and utilities
tqdm>=4.65.0

# For saving and loading models
h5py>=3.9.0
